Fixed some bad references and incorrect in game browser header usages.

+++ paplink/models.py
@@ -12,7 +12,7 @@ def get_sentinel_user():
 class Paplink(models.Model):
     papdatetime = models.DateTimeField(default=timezone.now())
     papdate = models.DateField(default=date.today())
-    duration = models.PositiveSmallIntegerField()
+    duration = models.PositiveIntegerField()
     fleet = models.CharField(max_length=254, default="")
     name = models.CharField(max_length=254, unique_for_date="papdate")
     creator = models.ForeignKey(User, on_delete=models.SET(get_sentinel_user))
